chore: rename wal.rs to single.rs
This better encompasses the singular wal nature which is used to build into a multi-wal

Consider better naming for this
